requirements.txt
daemonize
python-lirc
pyzmq